DOKY-237 Refactor
Simplify application structure and enhance configurations.

Removed unused `Mapper` class and streamlined the main application entry point with `runApplication`. Updated `application.properties` and configuration metadata to store all configuration used by application. Improved `README` with configuration setup instructions.
