--- src/main/kotlin/org/hkurh/doky/Mapper.kt
@@ -1,39 +1,0 @@
-/*
- * This file is part of the Doky Project.
- *
- * Copyright (C) 2005
- *  - Hanna Kurhuzenkava (hanna.kuehuzenkava@outlook.com)
- *  - Anton Kurhuzenkau (kurguzenkov@gmail.com)
- *
- * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty
- * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with this program. If not, see [Hyperlink removed
- * for security reasons]().
- *
- * Contact Information:
- *  - Project Homepage: https://github.com/hanna-eismant/doky
- */
-
-package org.hkurh.doky
-
-import org.hkurh.doky.documents.db.DocumentEntity
-import org.hkurh.doky.search.DocumentIndexData
-import java.time.ZoneOffset
-
-fun DocumentEntity.toIndexData(): DocumentIndexData {
-    val entity = this
-    return DocumentIndexData(
-        id = entity.id.toString(),
-        name = entity.name,
-        description = entity.description,
-        fileName = entity.fileName,
-        createdDate = entity.createdDate?.toInstant()?.atOffset(ZoneOffset.UTC),
-        modifiedDate = entity.modifiedDate?.toInstant()?.atOffset(ZoneOffset.UTC),
-        createdBy = entity.createdBy?.id.toString(),
-        modifiedBy = entity.modifiedBy?.id.toString()
-    )
-}
