+++ src/main/resources/application.properties
@@ -1,7 +1,6 @@
-# local port to start server
+# main configuration
 server.port=8080
 spring.application.name=doky
-#doky.app.host=
 
 # swagger
 springdoc.swagger-ui.path=/swagger-ui.html
@@ -14,29 +13,38 @@ springdoc.default-produces-media-type=application/json
 spring.session.jdbc.initialize-schema=never
 spring.jpa.open-in-view=false
 spring.flyway.locations=classpath:migration/sqlserver
-# logging
-server.error.include-stacktrace=never
-logging.level.web=WARN
-logging.level.org.hkurh.doky=DEBUG
-logging.level.org.hkurh.doky.security.JwtAuthorizationFilter=ERROR
+spring.datasource.password=
+spring.datasource.url=
+spring.datasource.username=
+
 # upload config
 spring.servlet.multipart.max-file-size=10MB
 spring.servlet.multipart.max-request-size=10MB
+
 # file storage
-#doky.filestorage.type=
-#doky.filestorage.path=
-#doky.filestorage.azure.connection=
-#doky.filestorage.azure.container=
+doky.filestorage.type=azure-blob
+doky.filestorage.path=./prod
+doky.filestorage.azure.connection=DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;
+doky.filestorage.azure.container=documents
+
 # security
 spring.security.user.name=user
 spring.security.user.password=<CHANGE_ME>
+
 # password
 doky.password.reset.token.duration=10
+
 # actuator
 management.endpoints.web.base-path=/api/actuator
 management.endpoints.web.exposure.include=*
 management.health.defaults.enabled=true
+
 # email
+doky.email.server.type=sendgrid
+doky.app.host=http://localhost:${server.port}
+doky.email.sender.email=doky-dev@outlook.com
+doky.email.sender.name=Doky Team
+# smtp configs
 spring.mail.host=smtp.sendgrid.net
 spring.mail.port=587
 spring.mail.username=apikey
@@ -44,19 +52,18 @@ spring.mail.password=<CHANGE_ME>
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
 spring.thymeleaf.check-template-location=false
-doky.email.sender.email=doky-dev@outlook.com
-doky.email.sender.name=Doky Team
 doky.email.templates.path=mail/templates/
-doky.email.server.type=sendgrid
+# sendgrid api configuration
+doky.email.sendgrid.api-key=<CHANGE_ME>
 doky.email.registration.sendgrid.template-id=d-a445db8eb8ac4d86ba38277ac2bb82da
 doky.email.registration.subject=Welcome to Doky
 doky.email.reset-password.sendgrid.template-id=d-9306f59a51b6457a950419b28525ab91
 doky.email.reset-password.subject=Doky Reset Password Request
 # kafka
-#doky.kafka.username=<CHANGE_ME>
-#doky.kafka.password=<CHANGE_ME>
+doky.kafka.username=<CHANGE_ME>
+doky.kafka.password=<CHANGE_ME>
 spring.kafka.properties.sasl.mechanism=PLAIN
-#spring.kafka.bootstrap-servers=<CHANGE_ME>
+spring.kafka.bootstrap-servers=<CHANGE_ME>
 spring.kafka.properties.sasl.jaas.config=org.apache.kafka.common.security.plain.PlainLoginModule required username='${doky.kafka.username}' password='${doky.kafka.password}';
 spring.kafka.properties.security.protocol=SASL_SSL
 spring.kafka.properties.session.timeout.ms=45000
@@ -67,3 +74,12 @@ doky.kafka.emails.group.id=doky-email-service
 doky.kafka.emails.consumer.id=email-notification-consumer
 doky.kafka.emails.consumer.autostart=true
 doky.kafka.emails.concurrency=1
+# search (Azure AI Search)
+azure.search.api-key=<CHANGE_ME>
+azure.search.endpoint=https://doky.search.windows.net
+azure.search.index-name=documents-local
+# logging
+server.error.include-stacktrace=never
+logging.level.web=WARN
+logging.level.org.hkurh.doky=DEBUG
+logging.level.org.hkurh.doky.security.JwtAuthorizationFilter=ERROR
